Add img directory in root, add basic structure to recipe page
Create an image directory in the root directory to store images for
use in other pages. Add an image for the first recipe.

Flesh out the first garlic shrimp recipe with headers, images, and lists
to provide the basic structure for all subsequent recipes. Provide alt
text for image as well.

Paragraph breaks will be necessary, as well as an image resize for the
poster, especially if subsequent images are varied sizes. Curriculum
has yet to cover this, so I will avoid touching it for now.